# NOXUS-Agent-API-v2/noxus-local-v2/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NOXUS API",
    "description": "JSON local + NoxusAgent + futura extensão Noxus VS Code.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/findings": {
      "post": {
        "tags": [
          "Ingestão"
        ],
        "summary": "Receive Agent Or Extension",
        "operationId": "receive_agent_or_extension_api_findings_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanEnvelope"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Findings",
        "operationId": "findings_api_findings_get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Id"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "tool",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tool"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/integrations/noxus-vscode/scans": {
      "post": {
        "tags": [
          "EXTENSÃO NOXUS VS CODE"
        ],
        "summary": "Receive Noxus Vscode Extension Scan",
        "operationId": "receive_noxus_vscode_extension_scan_api_integrations_noxus_vscode_scans_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanEnvelope"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/schema": {
      "get": {
        "summary": "Schema",
        "operationId": "schema_api_schema_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scans": {
      "get": {
        "summary": "Scans",
        "operationId": "scans_api_scans_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scans/{scan_id}": {
      "get": {
        "summary": "Scan",
        "operationId": "scan_api_scans__scan_id__get",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/findings/{finding_id}/status": {
      "patch": {
        "summary": "Triage",
        "operationId": "triage_api_findings__finding_id__status_patch",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Finding Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Triage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/findings/{finding_id}/history": {
      "get": {
        "summary": "History",
        "operationId": "history_api_findings__finding_id__history_get",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Finding Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assets": {
      "get": {
        "summary": "Assets",
        "operationId": "assets_api_assets_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard": {
      "get": {
        "summary": "Dashboard",
        "operationId": "dashboard_api_dashboard_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Asset": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Name"
          },
          "type": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Type",
            "default": "web-api"
          },
          "repository_url": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Repository Url"
          },
          "branch": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Branch",
            "default": "main"
          },
          "commit": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9a-fA-F]{7,64}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit"
          },
          "local_ip": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Ip"
          },
          "application_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Url"
          },
          "consumed_apis": {
            "items": {
              "$ref": "#/components/schemas/ConsumedAPI"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Consumed Apis"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "repository_url"
        ],
        "title": "Asset"
      },
      "ConsumedAPI": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Name"
          },
          "base_url": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Base Url"
          },
          "source": {
            "type": "string",
            "enum": [
              "manual",
              "detected"
            ],
            "title": "Source",
            "default": "manual"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "base_url"
        ],
        "title": "ConsumedAPI"
      },
      "Dependency": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Name"
          },
          "version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "ecosystem": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Ecosystem"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Dependency"
      },
      "Developer": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Name"
          },
          "role": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "team": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Team"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Developer"
      },
      "Finding": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 8000,
            "title": "Description",
            "default": ""
          },
          "rule_id": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Rule Id"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "critical",
                  "high",
                  "medium",
                  "low",
                  "info"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "cwe": {
            "items": {
              "type": "string",
              "pattern": "^CWE-\\d+$"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Cwe"
          },
          "cve": {
            "items": {
              "type": "string",
              "pattern": "^CVE-\\d{4}-\\d{4,}$"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Cve"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "dependency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Dependency"
              },
              {
                "type": "null"
              }
            ]
          },
          "recommendation": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8000
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "rule_id",
          "location"
        ],
        "title": "Finding"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Location": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "line": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          },
          "column": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Column"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Location"
      },
      "Scan": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tool": {
            "type": "string",
            "enum": [
              "semgrep",
              "gitleaks",
              "dependency-check",
              "nikto",
              "sonar",
              "dependency-reputation",
              "extension-reputation"
            ],
            "title": "Tool"
          },
          "tool_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Version"
          },
          "category": {
            "type": "string",
            "enum": [
              "SAST",
              "DAST",
              "SCA",
              "SECRET",
              "EXTENSION"
            ],
            "title": "Category"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "type": "string",
            "format": "date-time",
            "title": "Finished At"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "partial",
              "failed"
            ],
            "title": "Status"
          },
          "trigger": {
            "type": "string",
            "enum": [
              "manual",
              "startup",
              "file-change",
              "dependency-change",
              "commit",
              "app-online",
              "scheduled",
              "ide-save",
              "ide-change",
              "extension-change"
            ],
            "title": "Trigger",
            "default": "manual"
          },
          "error": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "tool",
          "category",
          "started_at",
          "finished_at",
          "status"
        ],
        "title": "Scan"
      },
      "ScanEnvelope": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "source": {
            "type": "string",
            "enum": [
              "noxus-agent",
              "noxus-vscode-extension"
            ],
            "title": "Source"
          },
          "developer": {
            "$ref": "#/components/schemas/Developer"
          },
          "asset": {
            "$ref": "#/components/schemas/Asset"
          },
          "scan": {
            "$ref": "#/components/schemas/Scan"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "maxItems": 20000,
            "title": "Findings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "developer",
          "asset",
          "scan",
          "findings"
        ],
        "title": "ScanEnvelope"
      },
      "Triage": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "open",
              "in_progress",
              "resolved",
              "false_positive",
              "accepted_risk"
            ],
            "title": "Status"
          },
          "actor": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Actor"
          },
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "actor",
          "reason"
        ],
        "title": "Triage"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}